+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## v0.0.2 - 2021/06/30
 
+-   Added `IRegisterOptions.overwrite` — Allows calls to `register` overwrite same-name search indices, instead of throwing an exception.
 -   Fixed NPM install.
 
 ## v0.0.1 - 2021/06/30